Agrega persistencia (cargado de mapas) y mejora deslizamiento - Permite al usuario elegir una archivo binario del sistema para reconstruir el mapa (funcionamiento base, sin validacion) - El mapa cargado opera exactamente igual que al momento que fue creado - Mejora función de DRAG en el canvas

+++ editor/asset_selector.cpp
@@ -5,7 +5,7 @@
 
 #include "ui_assetselector.h"
 
-AssetSelector::AssetSelector(const QHash<uint8_t, AssetData>& tiles, const QHash<uint8_t, AssetData>& colliders,
+AssetSelector::AssetSelector(const QHash<uint8_t, AssetData>& tiles, const QHash<uint16_t, AssetData>& colliders,
                              QWidget* parent):
         QTabWidget(parent), ui(new Ui::AssetSelector), tiles(tiles), colliders(colliders) {
